+++ src/entities/items/api/fetchDetailItem.js
