+++ src/entities/items/ui/itemList/BestItem.jsx
@@ -1,12 +1,12 @@
 import React from "react";
-import heart from "../../../shared/assets/heart.svg";
+import heart from "../../../../shared/assets/heart.svg";
 import { Link } from "react-router-dom";
 const BestItem = ({ item }) => {
   console.log(item);
   return (
     <div key={item.id} className="">
       <div className="mb-1">
-        <Link to={`/products/${item.id}`} className="w-full">
+        <Link to={`/items/${item.id}`} className="w-full">
           <img className="w-full h-48 rounded-lg" src={item.images[0]}></img>
         </Link>
       </div>
