+++ src/entities/items/ui/itemList/BestItems.jsx
@@ -1,8 +1,8 @@
 import React from "react";
-import { fetchItems } from "../../../entities/items/fetchItems";
+import { fetchItems } from "../../api/fetchItems";
 import BestItem from "./BestItem";
 import { useQuery } from "@tanstack/react-query";
-import useResponsiveDisplayCount from "../../../shared/hooks/useResponsiveDisplayCount";
+import useResponsiveDisplayCount from "../../../../shared/hooks/useResponsiveDisplayCount";
 
 const BestItems = () => {
   const { bestDisplayCount } = useResponsiveDisplayCount(); // 훅 사용
