+++ src/features/addInquiry/InquiryButton.jsx
@@ -1,0 +1,14 @@
+import React from "react";
+
+const InquiryButton = () => {
+  return (
+    <button
+      type="submit"
+      className={"w-14 h-10 border rounded-lg bg-slate-400 text-white "}
+    >
+      등록
+    </button>
+  );
+};
+
+export default InquiryButton;
