+++ src/pages/Items/index.jsx
@@ -1,5 +1,5 @@
-import AllItems from "../../features/itemList/allItem/AllItems";
-import BestItems from "../../features/itemList/bestItem/BestItems";
+import AllItems from "../../entities/items/ui/itemList/AllItems";
+import BestItems from "../../entities/items/ui/itemList/BestItems";
 
 const Items = () => {
   return (
